+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added trying to fix the predicted character offsets using an alignment algorithm
+
 ### Fixed
 - Specifying the spacy version to 2.2.4 in `setup.py`